# services/production-dashboard/openapi.yaml
openapi: 3.0.0
info:
  title: Production Dashboard API
  description: WebSocket + REST API for real-time production updates in PrintShop OS
  version: 1.0.0
  contact:
    name: PrintShop OS Team
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: ws://localhost:3000
    description: WebSocket server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [operator, supervisor, admin]
        email:
          type: string
    Order:
      type: object
      properties:
        id:
          type: string
        jobId:
          type: string
        customerId:
          type: string
        customerName:
          type: string
        status:
          type: string
          enum: [pending_approval, approved, in_screen, in_printing, in_curing, in_quality_check, completed, on_hold, cancelled]
        priority:
          type: integer
        service:
          type: string
          enum: [screen, dtg, embroidery]
        quantity:
          type: integer
        dueDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    QueueItem:
      type: object
      properties:
        order:
          $ref: '#/components/schemas/Order'
        position:
          type: integer
        estimatedStartTime:
          type: string
          format: date-time
        blockers:
          type: array
          items:
            type: string
    Machine:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [screen_press, dtg_printer, embroidery_machine, curing_station]
        status:
          type: string
          enum: [idle, running, down, maintenance]
        currentJobId:
          type: string
        utilizationPercent:
          type: number
        lastMaintenance:
          type: string
          format: date-time
        uptime:
          type: number
    Employee:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [clocked_in, clocked_out, break]
        currentJobId:
          type: string
        clockedInAt:
          type: string
          format: date-time
        productivity:
          type: number
        skillSet:
          type: array
          items:
            type: string
...
    RESTResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
        timestamp:
          type: string
          format: date-time
security:
  - bearerAuth: []
paths:
...
  /api/production/queue/reorder:
    post:
      tags:
        - Queue
      summary: Reorder production queue
      description: Reorder the production queue (supervisor and admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Queue reordered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RESTResponse'
        '403':
          description: Insufficient permissions
  /api/production/queue/{orderId}:
    get:
      tags:
        - Queue
      summary: Get order in queue
      description: Get a specific order from the production queue
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/RESTResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/QueueItem'
        '404':
          description: Order not found
...
  /api/production/resources/machines/{machineId}:
    get:
      tags:
        - Resources
      summary: Get specific machine
      description: Get details of a specific machine
      parameters:
        - name: machineId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Machine details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/RESTResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Machine'
        '404':
          description: Machine not found
  /api/production/resources/machines/{machineId}/allocate:
    post:
      tags:
        - Resources
      summary: Allocate machine to order
      description: Allocate a machine to a specific order (supervisor and admin only)
      parameters:
        - name: machineId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  type: string
      responses:
        '200':
          description: Machine allocated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/RESTResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Machine'
        '403':
          description: Insufficient permissions
...
  /api/production/resources/employees/{employeeId}:
    get:
      tags:
        - Resources
      summary: Get specific employee
      description: Get details of a specific employee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Employee details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/RESTResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Employee'
        '404':
          description: Employee not found
  /api/production/resources/employees/{employeeId}/assign:
    post:
      tags:
        - Resources
      summary: Assign employee to order
      description: Assign an employee to a specific order (supervisor and admin only)
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  type: string
      responses:
        '200':
          description: Employee assigned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/RESTResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Employee'
        '403':
          description: Insufficient permissions
  /api/production/orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order details
      description: Get detailed information about a specific order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/RESTResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
  /api/production/status:
    post:
      tags:
        - Orders
      summary: Update order status
      description: Update the status of an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  type: string
                status:
                  type: string
                  enum: [pending_approval, approved, in_screen, in_printing, in_curing, in_quality_check, completed, on_hold, cancelled]
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/RESTResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Order'
        '403':
          description: Insufficient permissions
...
  /api/production/connections:
    get:
      tags:
        - Admin
      summary: Get connected users
      description: Get list of currently connected users (admin only)
      responses:
        '200':
          description: Connected users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/RESTResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          count:
                            type: integer
                          users:
                            type: array
                            items:
                              $ref: '#/components/schemas/User'
        '403':
          description: Insufficient permissions